+++ Loader/JsonConfigurationLoader.php
@@ -15,6 +15,7 @@ namespace Linkin\Bundle\SwaggerResolverBundle\Loader;
 
 use EXSyst\Component\Swagger\Swagger;
 use Linkin\Bundle\SwaggerResolverBundle\Merger\OperationParameterMerger;
+use Symfony\Component\Routing\RouterInterface;
 
 /**
  * @author Viktor Linkin <adrenalinkin@gmail.com>
@@ -28,11 +29,12 @@ class JsonConfigurationLoader extends AbstractFileConfigurationLoader
 
     /**
      * @param OperationParameterMerger $parameterMerger
+     * @param RouterInterface $router
      * @param string $pathToFile
      */
-    public function __construct(OperationParameterMerger $parameterMerger, string $pathToFile)
+    public function __construct(OperationParameterMerger $parameterMerger, RouterInterface $router, string $pathToFile)
     {
-        parent::__construct($parameterMerger, $pathToFile);
+        parent::__construct($parameterMerger, $router, $pathToFile);
 
         $this->pathToFile = $pathToFile;
     }